Added : Ability to add our own projects

+++ src/pages/admin/page.tsx
@@ -1,12 +1,11 @@
 import Admin from '@/components/Admin';
-import { Button } from '@/components/ui/button';
+import WrongPage from '@/components/WrongPage';
 import { client } from '@/lib/supabase/supabase';
-import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react'
 
 const AdminPage = () => {
-    const router = useRouter();
     const [validUser, setValidUser] = useState(false);
+
     useEffect(() => {
         const checkSession = async () => {
             const { data: { session } } = await client.auth.getSession();
@@ -17,32 +16,21 @@ const AdminPage = () => {
             }
             return;
         }
-
+        client.auth.onAuthStateChange((event, session) => {
+            if (event === "SIGNED_OUT") {
+                localStorage.clear();
+            }
+        })
         checkSession();
     }, []);
-
-    const handleLogout = () => {
-        localStorage.clear();
-        router.push("/");
-    }
 
     return (
         <>
             {
                 validUser ? (
-                    <div>
-                        <Admin />
-                    </div>
+                    <Admin />
                 ) : (
-                    <div>
-                        <p>
-                            Oops!! You're not supposed to be here!!
-                        </p>
-                        <Button onClick={handleLogout}>
-                            Return To Home
-                        </Button>
-                    </div>
-
+                    <WrongPage />
                 )
             }
         </>
